# openapi.yaml
openapi: "3.1.0"
info:
  title: LeanStock API
  version: "2.0.0"
  description: |
    Multi-tenant inventory management system with atomic transfers,
    dead stock decay, predictive reordering, reservation system,
    and full supplier/purchase-order workflows.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://leanstock.onrender.com
    description: Production (Render.com)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /locations:
    get:
      tags: [Locations]
      summary: List active locations
      responses:
        "200":
          description: Location list
    post:
      tags: [Locations]
      summary: Create location (manager+)
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: "Main Warehouse"
              address: "123 Storage St"
              city: "Almaty"
      responses:
        "201":
          description: Created
  /locations/{location_id}:
    get:
      tags: [Locations]
      summary: Get location by ID
      parameters:
        - name: location_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location details
    delete:
      tags: [Locations]
      summary: Soft-delete location (manager+)
      parameters:
        - name: location_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated

  # ── Inventory ──────────────────────────────────────────────────────────────
  /inventory:
    get:
      tags: [Inventory]
      summary: List inventory (cursor pagination)
      parameters:
        - name: location_id
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Inventory items
  /inventory/set:
    post:
      tags: [Inventory]
      summary: Set inventory quantity (manager+)
      requestBody:
        required: true
        content:
          application/json:
            example:
              product_id: 1
              location_id: 1
              quantity: 100
      responses:
        "201":
          description: Inventory updated
  /inventory/transfer:
    post:
      tags: [Inventory]
      summary: Atomic transfer between locations (SELECT FOR UPDATE)
      requestBody:
        required: true
        content:
          application/json:
            example:
              product_id: 1
              from_location_id: 1
              to_location_id: 2
              quantity: 30
      responses:
        "200":
          description: Transfer completed
        "409":
          description: Insufficient stock
...
  /inventory/forecast:
    get:
      tags: [Inventory]
      summary: Predictive reorder suggestions (moving average)
      parameters:
        - name: location_id
          in: query
          schema:
            type: integer
        - name: days_window
          in: query
          schema:
            type: integer
            default: 30
        - name: reorder_threshold_days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Forecast and reorder suggestions
          content:
            application/json:
              example:
                count: 2
                reorder_threshold_days: 7
                items:
                  - product_id: 1
                    product_name: "Widget Pro"
                    sku: "WGT-001"
                    location_id: 1
                    current_quantity: 15
                    min_stock: 10
                    avg_daily_consumption: 2.5
                    days_until_stockout: 6.0
                    needs_reorder: true
                    suggested_reorder_qty: 20
                    analysis_window_days: 30
...